Fix incorrect type error when trying to set stencil properties

+++ Packages/com.deltation.toon-rp/Editor/ShaderGUI/ToonRpShaderGuiBase.cs
@@ -328,17 +328,17 @@ namespace DELTation.ToonRP.Editor.ShaderGUI
                 if (enableStencil)
                 {
                     byte reference = ((StencilLayer) stencilPreset).ToReference();
-                    m.SetFloat(ForwardStencilRefId, reference);
-                    m.SetFloat(ForwardStencilReadMaskId, 255);
-                    m.SetFloat(ForwardStencilWriteMaskId, reference);
+                    m.SetInteger(ForwardStencilRefId, reference);
+                    m.SetInteger(ForwardStencilReadMaskId, 255);
+                    m.SetInteger(ForwardStencilWriteMaskId, reference);
                     m.SetFloat(ForwardStencilCompId, (int) CompareFunction.Always);
                     m.SetFloat(ForwardStencilPassId, (int) StencilOp.Replace);
                 }
                 else
                 {
-                    m.SetFloat(ForwardStencilRefId, 0);
-                    m.SetFloat(ForwardStencilReadMaskId, 255);
-                    m.SetFloat(ForwardStencilWriteMaskId, 255);
+                    m.SetInteger(ForwardStencilRefId, 0);
+                    m.SetInteger(ForwardStencilReadMaskId, 255);
+                    m.SetInteger(ForwardStencilWriteMaskId, 255);
                     m.SetFloat(ForwardStencilCompId, (int) CompareFunction.Always);
                     m.SetFloat(ForwardStencilPassId, (int) StencilOp.Keep);
                 }
